move validate cidr
